Change latitude and longitude from Float to Double so there is no need to cast them when used in LatLng.

+++ composeApp/src/androidMain/kotlin/charly/baquero/pocketmap/ui/map/MapComponent.android.kt
@@ -29,8 +29,8 @@ actual fun MapComponent(locationList: List<Location>?) {
                 Marker(
                     state = rememberUpdatedMarkerState(
                         position = LatLng(
-                            it.latitude.toDouble(),
-                            it.longitude.toDouble()
+                            it.latitude,
+                            it.longitude
                         )
                     ),
                     title = it.title,
